Ai2c Apps/Griffin-backend-dev/requirements.txt
Django==4.2.25
mssql-django==1.6
python-dotenv==1.1.0
openpyxl==3.1.5
pandas==2.2.3
djangorestframework==3.16.0
reportlab==4.4.1
azure-storage-blob==12.25.1
azure-monitor-opentelemetry==1.6.10
numpy==2.2.6
django-simple-history==3.8.0
django-ninja==1.4.2
django-cors-headers==4.7.0
django-polymorphic==4.1.0
unittest-xml-reporting==3.2.0
scikit-learn==1.6.1
gunicorn==23.0.0
django-csp==4.0
defusedcsv==2.0.0
defusedxml==0.7.1
azure-identity==1.23.0
pydantic==2.11.10